+++ django/applications/catmaid/static/js/widgets/overlay.js
@@ -15,16 +15,6 @@
  * Contains the current state of skeleton annotations.
  */
 var SkeletonAnnotations = {
-  // Colors that a node can take
-  atn_fillcolor : "rgb(0, 255, 0)",
-  active_skeleton_color: "rgb(255,255,0)",
-  active_skeleton_color_virtual: "rgb(255,255,0)",
-  inactive_skeleton_color: "rgb(255,0,255)",
-  inactive_skeleton_color_virtual: "rgb(255,0,255)",
-  inactive_skeleton_color_above: "rgb(0,0,255)",
-  inactive_skeleton_color_below: "rgb(255,0,0)",
-  root_node_color: "rgb(255,0,0)",
-  leaf_node_color: "rgb(128,0,0)",
 
   /**
    * Data of the active Treenode or ConnectorNode. Its position is stored in
@@ -277,13 +267,6 @@ SkeletonAnnotations.getActiveNodeType = function() {
 
 SkeletonAnnotations.getActiveNodeSubType = function() {
   return this.atn.subtype;
-};
-
-/**
- * Get the fill color for an active node.
- */
-SkeletonAnnotations.getActiveNodeColor = function() {
-  return this.atn_fillcolor;
 };
 
 /**
@@ -731,6 +714,34 @@ SkeletonAnnotations.TracingOverlay.Settings = new CATMAID.Settings(
           },
           scale: {
             default: 1.0
+          },
+          // Colors that a node can take
+          active_node_color : {
+            default: "rgb(0,255,0)",
+          },
+          active_skeleton_color: {
+            default: "rgb(255,255,0)",
+          },
+          active_skeleton_color_virtual: {
+            default: "rgb(255,255,0)",
+          },
+          inactive_skeleton_color: {
+            default: "rgb(255,0,255)",
+          },
+          inactive_skeleton_color_virtual: {
+            default: "rgb(255,0,255)",
+          },
+          inactive_skeleton_color_above: {
+            default: "rgb(0,0,255)",
+          },
+          inactive_skeleton_color_below: {
+            default: "rgb(255,0,0)",
+          },
+          root_node_color: {
+            default: "rgb(255,0,0)",
+          },
+          leaf_node_color: {
+            default: "rgb(128,0,0)",
           }
         },
         migrations: {}